Rename helper module and move into a folder
Having the helper module as a file and not a `mod.rs` in a folder caused
rust to give warnings about unused functions.
